add expoImage component
